Faster granule detection via FFT-based blurring (#3)
Adds a `use_fft` keyword to the configuration file that allows for a FFT based approach to the Gaussian blurring in the DoG step for finding granules. This defaults to `False`.

This creates a new `GranuleDetectorFFT` class that overrides the `labelGranules` function, and some setup code before the first frame in `process_image`, so all old plotting/code should be unaffected.

Other minor changes:
- Falls back to the "bioformats" image opener with a log.warning if we don't recognise the format. 
- Added a `.gitignore` for `.pyc` and `__pycache__` files. 
- Deleted some variables that were defined in the `process_image` file that went unused.
- Made the frame/detected granule plotting prettier.
- `mkdir` the plotting directories (only if the `out` directory exists) to save on crashes mid-run.